Handles polars upgrade to 1.0.0
Decisions:
1. Not backwards compatible -- we're effectively wrapping the library
   and we don't want to provide a bridge. This is polars versioning, not
   Hamilton versioning, however, so it's OK to keep this a patch bump
   (any polars code would have broken already).
2. They deprecated the usage of a connection string (sqllite://) in the
   PolarsDatabase{reader/writer} -- we've deprecated/changed the type as
   well
3. It switches based on polars version so people will only get the error
   for new polars upgrades

Otherwise, longer term, we're going to want to fix this -- moving away
from these blind wrappers.
